The Hash Tag will be the example of what box-sizing should be placed and the margin and padding will be at a 0%.
The body will have a line height so there can be evenly space between the words and images.

the header will be the title so the title will always have a text-align center and padding.

Nav will be placed on the left thus float place whatever is inside of the nav to go left and it will add width and padding and background-color.

Main will also be at the float left but the width and padding will have a space evenly so it wont get mixed up.

aside will also be at the left but the width will always have the % and padding and margin does not have to have to add %.

But then again the width and height does give the image its good size so it wont be too big nor too small.